pyproject.toml
[project]
name = "pehraz-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.2",
    "django-cors-headers>=4.9.0",
    "django-ninja>=1.5.3",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.53.1",
]
